GOTOOLS-287 Allow passing `exclude` to `Analyzer`
## Context

Jira Issue: https://datadoghq.atlassian.net/browse/GOTOOLS-287

We want to be able to pass `exclude` to the `Analyzer` so we can exclude
some functions/methods from `errcheck`. This doesn't really work when
using `rules_go`'s `nogo`, because `errcheck` requires a file to read
for `exclude`, and `nogo` doesn't support passing files to analyzers:
https://github.com/bazel-contrib/rules_go/issues/3234. Similarly:
https://github.com/kisielk/errcheck/issues/243

We're working around this by adding a helper for creating a new
`Analyzer` that takes a `map[string]bool` of functions/methods to
exclude. This lets us wrap the `errcheck` analyzer and pass in the
excludes from Go.

Hopefully the `nogo` issue gets resolved at some point, and we can rely
on the original interface of `errcheck`. Until that point, this change
should allow us to work around the incompatibilities between the two.